Extract document consolidated package
