+++ src/main/java/com/code4ro/legalconsultation/converters/DocumentConsolidatedMapper.java
@@ -2,8 +2,8 @@ package com.code4ro.legalconsultation.converters;
 
 import com.code4ro.legalconsultation.document.metadata.mapper.DocumentMetadataMapper;
 import com.code4ro.legalconsultation.document.node.mapper.DocumentNodeMapper;
-import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.dto.DocumentConsolidatedDto;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.user.mapper.UserMapper;
 import org.mapstruct.Mapper;
