+++ src/main/java/com/code4ro/legalconsultation/document/configuration/model/persistence/DocumentConfiguration.java
@@ -1,5 +1,6 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.document.configuration.model.persistence;
 
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
