+++ src/main/java/com/code4ro/legalconsultation/document/configuration/repository/DocumentConfigurationRepository.java
@@ -1,6 +1,6 @@
-package com.code4ro.legalconsultation.repository;
+package com.code4ro.legalconsultation.document.configuration.repository;
 
-import com.code4ro.legalconsultation.model.persistence.DocumentConfiguration;
+import com.code4ro.legalconsultation.document.configuration.model.persistence.DocumentConfiguration;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
