+++ src/main/java/com/code4ro/legalconsultation/document/consolidated/model/dto/DocumentConsolidatedDto.java
@@ -1,7 +1,9 @@
-package com.code4ro.legalconsultation.model.dto;
+package com.code4ro.legalconsultation.document.consolidated.model.dto;
 
+import com.code4ro.legalconsultation.document.configuration.model.dto.DocumentConfigurationDto;
 import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
 import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeDto;
+import com.code4ro.legalconsultation.model.dto.BaseEntityDto;
 import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import lombok.Getter;
 import lombok.Setter;
