+++ src/main/java/com/code4ro/legalconsultation/document/consolidated/model/persistence/DocumentConsolidated.java
@@ -1,7 +1,9 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.document.consolidated.model.persistence;
 
+import com.code4ro.legalconsultation.document.configuration.model.persistence.DocumentConfiguration;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import lombok.Getter;
 import lombok.Setter;
