+++ src/main/java/com/code4ro/legalconsultation/document/core/controller/DocumentController.java
@@ -1,12 +1,14 @@
-package com.code4ro.legalconsultation.controller;
+package com.code4ro.legalconsultation.document.core.controller;
 
+import com.code4ro.legalconsultation.document.consolidated.model.dto.DocumentConsolidatedDto;
+import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentViewDto;
 import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
 import com.code4ro.legalconsultation.model.DocumentUserAssignmentDto;
 import com.code4ro.legalconsultation.model.dto.*;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.model.persistence.DocumentExportFormat;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
-import com.code4ro.legalconsultation.service.api.DocumentService;
+import com.code4ro.legalconsultation.document.core.service.DocumentService;
 import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
