+++ src/main/java/com/code4ro/legalconsultation/document/metadata/mapper/DocumentMetadataMapper.java
@@ -1,7 +1,7 @@
 package com.code4ro.legalconsultation.document.metadata.mapper;
 
 import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
-import com.code4ro.legalconsultation.model.dto.DocumentViewDto;
+import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentViewDto;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import org.mapstruct.Mapper;
