+++ src/main/java/com/code4ro/legalconsultation/document/metadata/model/dto/DocumentMetadataDto.java
@@ -1,7 +1,7 @@
 package com.code4ro.legalconsultation.document.metadata.model.dto;
 
 import com.code4ro.legalconsultation.model.dto.BaseEntityDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentType;
+import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentType;
 import lombok.Getter;
 import lombok.Setter;
