+++ src/main/java/com/code4ro/legalconsultation/document/metadata/model/persistence/DocumentType.java
@@ -1,0 +1,7 @@
+package com.code4ro.legalconsultation.document.metadata.model.persistence;
+
+public enum DocumentType {
+    HG,
+    OUG,
+    LEGE
+}
