+++ src/main/java/com/code4ro/legalconsultation/document/metadata/service/DocumentMetadataService.java
@@ -2,7 +2,7 @@ package com.code4ro.legalconsultation.document.metadata.service;
 
 import com.code4ro.legalconsultation.document.metadata.mapper.DocumentMetadataMapper;
 import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
-import com.code4ro.legalconsultation.model.dto.DocumentViewDto;
+import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentViewDto;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.document.metadata.repository.DocumentMetadataRepository;
 import org.springframework.beans.factory.annotation.Autowired;
