--- src/main/java/com/code4ro/legalconsultation/model/persistence/DocumentType.java
@@ -1,7 +1,0 @@
-package com.code4ro.legalconsultation.model.persistence;
-
-public enum DocumentType {
-    HG,
-    OUG,
-    LEGE
-}
