+++ src/main/java/com/code4ro/legalconsultation/service/export/DocumentExporter.java
@@ -1,6 +1,6 @@
 package com.code4ro.legalconsultation.service.export;
 
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 
 public interface DocumentExporter {
     byte[] export(DocumentConsolidated document);
