+++ src/main/java/com/code4ro/legalconsultation/service/export/PDFExporter.java
@@ -1,7 +1,7 @@
 package com.code4ro.legalconsultation.service.export;
 
 import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
 import com.lowagie.text.DocumentException;
 import com.lowagie.text.pdf.BaseFont;
