+++ src/main/java/com/code4ro/legalconsultation/service/impl/pdf/PDFServiceImpl.java
@@ -2,7 +2,7 @@ package com.code4ro.legalconsultation.service.impl.pdf;
 
 import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.converters.PdfHandleMapper;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.model.persistence.PdfHandle;
 import com.code4ro.legalconsultation.repository.PdfHandleRepository;
 import com.code4ro.legalconsultation.service.api.PDFService;
