+++ src/test/java/com/code4ro/legalconsultation/document/consolidated/service/DocumentConsolidatedServiceTest.java
@@ -1,13 +1,13 @@
-package com.code4ro.legalconsultation.service;
+package com.code4ro.legalconsultation.document.consolidated.service;
 
 import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.converters.DocumentConsolidatedMapper;
-import com.code4ro.legalconsultation.model.persistence.DocumentConfiguration;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.configuration.model.persistence.DocumentConfiguration;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
-import com.code4ro.legalconsultation.repository.DocumentConsolidatedRepository;
-import com.code4ro.legalconsultation.service.impl.DocumentConsolidatedService;
+import com.code4ro.legalconsultation.document.consolidated.repository.DocumentConsolidatedRepository;
+import com.code4ro.legalconsultation.document.consolidated.service.impl.DocumentConsolidatedService;
 import com.code4ro.legalconsultation.document.node.factory.DocumentNodeFactory;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import org.junit.Test;
