+++ src/test/java/com/code4ro/legalconsultation/document/core/controller/DocumentControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.code4ro.legalconsultation.controller;
+package com.code4ro.legalconsultation.document.core.controller;
 
 import com.amazonaws.util.json.Jackson;
 import com.code4ro.legalconsultation.common.controller.AbstractControllerIntegrationTest;
@@ -6,12 +6,12 @@ import com.code4ro.legalconsultation.comment.factory.CommentFactory;
 import com.code4ro.legalconsultation.document.node.factory.DocumentNodeFactory;
 import com.code4ro.legalconsultation.factory.PdfFileFactory;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
-import com.code4ro.legalconsultation.model.dto.DocumentViewDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentConfiguration;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentViewDto;
+import com.code4ro.legalconsultation.document.configuration.model.persistence.DocumentConfiguration;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
-import com.code4ro.legalconsultation.repository.DocumentConsolidatedRepository;
+import com.code4ro.legalconsultation.document.consolidated.repository.DocumentConsolidatedRepository;
 import com.code4ro.legalconsultation.document.metadata.repository.DocumentMetadataRepository;
 import com.code4ro.legalconsultation.document.node.repository.DocumentNodeRepository;
 import org.junit.Test;
