+++ src/test/java/com/code4ro/legalconsultation/service/DocumentServiceTest.java
@@ -3,9 +3,9 @@ package com.code4ro.legalconsultation.service;
 import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.converters.DocumentConsolidatedMapper;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
-import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
+import com.code4ro.legalconsultation.document.consolidated.model.dto.DocumentConsolidatedDto;
 import com.code4ro.legalconsultation.document.metadata.model.dto.DocumentMetadataDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.document.consolidated.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.document.metadata.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.model.persistence.DocumentExportFormat;
 import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
@@ -13,9 +13,9 @@ import com.code4ro.legalconsultation.model.persistence.PdfHandle;
 import com.code4ro.legalconsultation.service.api.MailApi;
 import com.code4ro.legalconsultation.service.export.DocumentExporterFactory;
 import com.code4ro.legalconsultation.service.export.PDFExporter;
-import com.code4ro.legalconsultation.service.impl.DocumentConsolidatedService;
+import com.code4ro.legalconsultation.document.consolidated.service.impl.DocumentConsolidatedService;
 import com.code4ro.legalconsultation.document.metadata.service.DocumentMetadataService;
-import com.code4ro.legalconsultation.service.impl.DocumentServiceImpl;
+import com.code4ro.legalconsultation.document.core.service.impl.DocumentServiceImpl;
 import com.code4ro.legalconsultation.service.impl.pdf.PDFServiceImpl;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.user.service.UserService;
